landmark path working

+++ edge.h
@@ -53,7 +53,7 @@ class Edge
      * @param w - the weight of the edge
      * @param lbl - the edge label
      */
-    Edge(Vertex u, Vertex v, int w, string lbl)
+    Edge(Vertex u, Vertex v, double w, string lbl)
         : source(u), dest(v), label(lbl), weight(w)
     { /* nothing */
     }
@@ -87,7 +87,7 @@ class Edge
     /**
      * Gets edge weight.
      */
-    int getWeight() const
+    double getWeight() const
     {
         return this->weight;
     }
@@ -111,6 +111,6 @@ class Edge
     
 private:
     string label; /**< The edge label **/
-    int weight; /**< The edge weight (if in a weighed graph) **/
+    double weight; /**< The edge weight (if in a weighed graph) **/
 
 };
